State ai implemented + some other stuff

+++ Interactive Board/Assets/Scripts/Asteroids Battle Royale/ABR_Ship.cs
@@ -9,7 +9,7 @@ public class ABR_Ship : MonoBehaviour
 
     private float m_thrustMult = 1;
     private bool m_doThrust = false;
-    private int m_turn = 0;
+    private float m_turn = 0;
 
     private float m_thrustPower = 10.0f;
     private float m_turnPower = 240.0f;
@@ -18,9 +18,8 @@ public class ABR_Ship : MonoBehaviour
 
     public float ThrustMult { get { return m_thrustMult; } set { m_thrustMult = value; } }
     public bool DoThrust { get { return m_doThrust; } set { m_doThrust = value; } }
-    public int Turn { get { return m_turn; } set { m_turn = Mathf.Clamp(value, -1, 1); } }
-
-
+    public float Turn { get { return m_turn; } set { m_turn = Mathf.Clamp(value, -1, 1); } }
+    
     // Start is called before the first frame update
     void Start()
     {
